+++ src/modules/admin/data/TenantActiveDataProvider.php
@@ -4,6 +4,8 @@ namespace davidhirtz\yii2\tenant\modules\admin\data;
 
 use davidhirtz\yii2\skeleton\data\ActiveDataProvider;
 use davidhirtz\yii2\tenant\models\Tenant;
+use yii\data\Pagination;
+use yii\data\Sort;
 
 class TenantActiveDataProvider extends ActiveDataProvider
 {
@@ -26,14 +28,34 @@ class TenantActiveDataProvider extends ActiveDataProvider
         if ($this->searchString !== null) {
             $this->query->andFilterWhere(['like', 'name', $this->searchString]);
         }
+
+        $this->query->orderBy(['position' => SORT_ASC]);
+    }
+
+    public function getPagination(): Pagination|false
+    {
+        return !$this->isOrderedByPosition() ? parent::getPagination() : false;
+    }
+
+    public function getSort(): Sort|false
+    {
+        return !$this->isOrderedByPosition() ? parent::getSort() : false;
     }
 
     public function setSort($value): void
     {
         if (is_array($value)) {
-            $value['defaultOrder'] ??= ['updated_at' => SORT_DESC];
+            $value['defaultOrder'] ??= ['position' => SORT_ASC];
         }
 
         parent::setSort($value);
+    }
+
+    public function isOrderedByPosition(): bool
+    {
+        return in_array(key($this->query->orderBy ?? []), [
+            Tenant::tableName() . '.[[position]]',
+            'position',
+        ]);
     }
 }